% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racespeed.R
\name{racetime}
\alias{racetime}
\title{Calculate race time}
\usage{
racetime(fart)
}
\arguments{
\item{fart}{Speed in km/h
Example racetime(24)}
}
\description{
Add distance and time in seconds to get speed in km/h
}
