% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcalc.R
\name{runspeed}
\alias{runspeed}
\title{Calculate running speed}
\usage{
runspeed(distanse, tid)
}
\arguments{
\item{distanse}{Distance in meters.}

\item{tid}{Time in seconds.
Example runspeed(400, 60)}
}
\description{
Add distance and time in seconds to get speed in km/h
}
