% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splittimes.R
\name{splittimes}
\alias{splittimes}
\title{Calculate split times}
\usage{
splittimes(distanse, tid)
}
\arguments{
\item{distanse}{Distance in meters.}

\item{tid}{Time in seconds.
Example splitimes(400, 60)}
}
\description{
Add distance and time in seconds to get splitimes per 100 meter in seconds
}
